moving file path generator and enum classes to packages
